requirements.txt
typing-extensions
python-telegram-bot
moviepy==1.0.3
imageio[ffmpeg]
